requirements.txt
# aSDLC Python Dependencies
# See docs/System_Design.md for architecture details

# Core dependencies
redis>=5.0.0
pydantic>=2.0.0

# Async support
anyio>=4.0.0

# Knowledge Store (RAG)
elasticsearch[async]>=8.0.0,<9.0.0
sentence-transformers>=2.2.0

# AST Parsing (Repo Mapper)
tree-sitter>=0.25.0
tree-sitter-typescript>=0.23.0
tree-sitter-javascript>=0.25.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
testcontainers[postgres]>=3.7.0

# Code quality
ruff>=0.1.0
mypy>=1.5.0

# Type stubs
types-redis>=4.6.0

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Metrics collection
prometheus_client>=0.19.0
psutil>=5.9.0

# HTTP client for VictoriaMetrics proxy
httpx>=0.25.0

# Kubernetes client
kubernetes>=28.0.0

# Encryption for API keys
cryptography>=42.0.0

# LLM Provider SDKs
google-generativeai>=0.8.0
anthropic>=0.40.0
openai>=1.50.0

# PostgreSQL persistence layer
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.0

# Development
python-dotenv>=1.0.0
slack-sdk==3.33.4
